+++ tests/test_rl.py
@@ -20,7 +20,7 @@ from pydrobert.torch.modules import TimeDistributedReturn
 
 @pytest.mark.parametrize("batch_first", [True, False])
 @pytest.mark.parametrize("gamma", [0.0, 0.95])
-def test_time_distributed_return(device, batch_first, gamma):
+def test_time_distributed_return(device, batch_first, gamma, jit_type):
     steps, batch_size = 1000, 30
     r = torch.randn(steps, batch_size, device=device)
     exp = torch.empty_like(r)
@@ -31,5 +31,11 @@ def test_time_distributed_return(device, batch_first, gamma):
         r = r.t().contiguous()
         exp = exp.t().contiguous()
     time_distributed_return = TimeDistributedReturn(gamma, batch_first)
+    if jit_type == "script":
+        time_distributed_return = torch.jit.script(time_distributed_return)
+    elif jit_type == "trace":
+        time_distributed_return = torch.jit.trace(
+            time_distributed_return, (torch.empty(1, 1),)
+        )
     act = time_distributed_return(r)
     assert torch.allclose(exp, act, atol=1e-5)